feat: 보드 유저 확인
1. 보드 생성 시에 유저 추가
2. 보드 수정, 삭제 시에 로그인한 유저 확인 추가
3. 보드 수정, 삭제 시에 return 값 Void로 변경

+++ src/main/java/com/example/gamemate/domain/board/service/BoardService.java
@@ -11,12 +11,16 @@ import com.example.gamemate.domain.board.repository.BoardRepository;
 import com.example.gamemate.domain.comment.dto.CommentFindResponseDto;
 import com.example.gamemate.domain.comment.entity.Comment;
 import com.example.gamemate.domain.comment.repository.CommentRepository;
+import com.example.gamemate.domain.user.entity.User;
+import com.example.gamemate.global.config.auth.CustomUserDetails;
 import com.example.gamemate.global.exception.ApiException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +28,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static com.example.gamemate.global.constant.ErrorCode.BOARD_NOT_FOUND;
+import static com.example.gamemate.global.constant.ErrorCode.FORBIDDEN;
 
 @Service
 @RequiredArgsConstructor
@@ -38,15 +43,18 @@ public class BoardService {
      * @return
      */
     @Transactional
-    public BoardResponseDto createBoard(BoardRequestDto dto) {
+    public BoardResponseDto createBoard(User loginUser, BoardRequestDto dto) {
         // 게시글 생성
-        Board newBoard = new Board(dto.getCategory(),dto.getTitle(),dto.getContent());
+        Board newBoard = new Board(dto.getCategory(),dto.getTitle(),dto.getContent(), loginUser);
         Board createdBoard = boardRepository.save(newBoard);
         return new BoardResponseDto(
                 createdBoard.getBoardId(),
                 createdBoard.getCategory(),
                 createdBoard.getTitle(),
-                createdBoard.getContent()
+                createdBoard.getContent(),
+                createdBoard.getUser().getNickname(),
+                createdBoard.getCreatedAt(),
+                createdBoard.getModifiedAt()
         );
     }
 
@@ -117,13 +125,18 @@ public class BoardService {
      * @return
      */
     @Transactional
-    public void updateBoard(Long id, BoardRequestDto dto) {
+    public void updateBoard(User loginUser, Long id, BoardRequestDto dto) {
         // 게시글 조회
         Board findBoard = boardRepository.findById(id)
                 .orElseThrow(()->new ApiException(BOARD_NOT_FOUND));
 
+        // 게시글 작성자와 로그인한 사용자 확인
+        if(!findBoard.getUser().getId().equals(loginUser.getId())) {
+            throw new ApiException(FORBIDDEN);
+        }
+
         findBoard.updateBoard(dto.getCategory(),dto.getTitle(),dto.getContent());
-        Board updatedBoard = boardRepository.save(findBoard);
+        boardRepository.save(findBoard);
     }
 
     /**
@@ -131,10 +144,15 @@ public class BoardService {
      * @param id
      */
     @Transactional
-    public void deleteBoard(Long id) {
+    public void deleteBoard(User loginUser, Long id) {
         //게시글 조회
         Board findBoard = boardRepository.findById(id)
                 .orElseThrow(()->new ApiException(BOARD_NOT_FOUND));
+
+        // 게시글 작성자와 로그인한 사용자 확인
+        if(!findBoard.getUser().getId().equals(loginUser.getId())) {
+            throw new ApiException(FORBIDDEN);
+        }
 
         boardRepository.delete(findBoard);
     }
